requirements.txt
bs4
lxml
python-multipart
fastapi
uvicorn
requests
azure-storage-blob